visualization/requirements.txt
# GRPC Python setup requirements
protobuf>=3.5.0
grpcio-tools
gunicorn
flask
pytz
